+++ pyxel/calibration/fitting_datatree.py
@@ -325,6 +325,12 @@ class ModelFittingDataTree(ProblemSingleObjective):
         if not isinstance(simulated_data, xr.DataArray):
             raise TypeError("Expected a 'DataArray'")
 
+        if "y" not in simulated_data.dims or "x" not in simulated_data.dims:
+            raise ValueError(
+                f"Missing dimensions 'y' or 'x' in container '{self.sim_output:s}'. "
+                f"{simulated_data=}"
+            )
+
         if self.sim_fit_range is not None:
             simulated_data = simulated_data.isel(indexers=self.sim_fit_range.to_dict())
 
@@ -441,10 +447,17 @@ class ModelFittingDataTree(ProblemSingleObjective):
                     weighting=weighting,  # TODO: 'weighting' should be a 'DataArray'
                 )
 
-        except Exception:
+        except Exception as exc:
             logging.exception(
-                "Catch an exception in 'fitness' for ModelFitting: %r.", self
+                "Catch an exception in 'fitness' for ModelFitting: %r with decision vector: %r.",
+                self,
+                decision_vector_1d,
             )
+            print(f"--- 3 - {exc=}, {exc.__notes__}")
+            # logging.warning('--- 3.1')
+            # logging.error('--- 3.1')
+            # logging.handlers[0].flush()
+
             raise
 
         return [overall_fitness]
@@ -477,7 +490,7 @@ class ModelFittingDataTree(ProblemSingleObjective):
 
         return parameters
 
-    def apply_parameters(
+    def _apply_parameters(
         self, processor: Processor, parameter: np.ndarray
     ) -> "DataTree":
         """Create a new ``Processor`` with new parameters."""
@@ -496,17 +509,19 @@ class ModelFittingDataTree(ProblemSingleObjective):
         self, parameters: "xr.DataArray"
     ) -> pd.DataFrame:
         """TBW."""
-        assert "island" in parameters.dims
-        assert "param_id" in parameters.dims
+        if "island" not in parameters.dims:
+            raise KeyError("Missing dimension 'island'")
+        if "param_id" not in parameters.dims:
+            raise KeyError("Missing dimension 'param_id'")
 
         lst = []
         for id_processor, processor in enumerate(self.param_processor_list):
             delayed_processor = delayed(processor)
 
             for idx_island, params_array in parameters.groupby("island"):
-                params: np.ndarray = params_array.data  # type
+                params: np.ndarray = params_array.to_numpy()
 
-                result_datatree: DataTree = delayed(self.apply_parameters)(
+                result_datatree: DataTree = delayed(self._apply_parameters)(
                     processor=delayed_processor, parameter=params
                 )